replace numeric to integer

+++ inital_setup/script.sql
@@ -1,6 +1,6 @@
 create table states
 (
-    state_id  numeric primary key,
+    state_id  integer primary key,
     name      varchar(256) not null,
     longitude float        not null,
     latitude  float        not null
@@ -11,14 +11,14 @@ create unique index state_idx
 
 create table info_per_day
 (
-    info_per_day_id     numeric primary key,
-    state_id            numeric not null,
-    confirmed           numeric,
-    deaths              numeric,
-    recovered           numeric,
-    active              numeric,
-    people_tested       numeric,
-    people_hospitalized numeric,
+    info_per_day_id     integer primary key,
+    state_id            integer not null,
+    confirmed           integer,
+    deaths              integer,
+    recovered           integer,
+    active              integer,
+    people_tested       integer,
+    people_hospitalized integer,
     date                date,
     foreign key (state_id) references states (state_id)
 );
